+++ packages/frontity/src/utils/decode/types.ts
@@ -1,5 +1,3 @@
-export * from "@frontity/types";
-
 export interface Decode {
   (text: string): string;
 }